+++ docs/setup/changelog.rst
@@ -17,6 +17,7 @@ Upcoming Changes (main branch)
 - ``array_equiv`` function.
 - Conversion functions for ``Plane``, ``Quaternion``, ``Projection`` and ``Basis``.
 - ``NDArray`` functions for getting slices as variant types (if shape is compatible).
+- ``complex64`` and ``complex128`` array creation shorthand functions.
 
 **Changed**